vagrant settings

+++ db/migrate/20190201080135_create_perceptrons.rb
@@ -2,7 +2,7 @@ class CreatePerceptrons < ActiveRecord::Migration[6.0]
   def change
     create_table :perceptrons do |t|
       t.integer :size
-      t.decimal :coefficient, precision: 1, scale: 8
+      t.float :coefficient, precision: 1, scale: 8
 
       t.timestamps
     end
